tasks: move `LinkObjects` and `LinkFunction` to `utils/db` package
